refactor(computer-use): extract window geometry
